recipes-toolbox/networkd-config/*.network: Use MAC for ClientIdentifier
In order to prevent the device from announcing via DHCP as a completely
difference device based on a randomized DUID/Machine-ID, set networkd to
announce with the MAC as the client identifier. This means regardless of
whether the system preserves its state across boots the device will
appear to be the same to the DHCP server. Relying on the MAC address to
provide system uniqueness.