rust-simple-init: Add initial support for DeviceManagerService
This service handles device events (uevent) and performs udev like
behaviour to autoload kernel modules for added devices (and at start).

The kmod module also adds support for loading the modules.alias and
performing the associated pattern matching to map modalias information
to kernel modules that need to be loaded.